requirements.txt
crewai
crewai-tools
boto3
python-dotenv
